requirements.txt
# Azure DevOps MCP Server Requirements
fastmcp>=0.1.0
pydantic>=2.0.0
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.7.22
python-dotenv>=1.0.0